Struct xi_rope::compare::RopeScanner

pub struct RopeScanner<'a> { /* fields omitted */ }

Utility for efficiently comparing two ropes.

Implementations

impl<'a> RopeScanner<'a>

pub fn new(base: &'a Rope, target: &'a Rope) -> Self

pub fn find_ne_char_back<T>(
&mut self,
base_off: usize,
targ_off: usize,
stop: T
) -> usize where
T: Into<Option<usize>>,

Starting from the two provided offsets in the corresponding ropes, Returns the distance, moving backwards, to the first non-equal codepoint. If no such position exists, returns the distance to the closest 0 offset.

if stop is not None, the scan will stop at if it reaches this value.

Examples

let one = Rope::from("hiii");
let two = Rope::from("siii");
let mut scanner = RopeScanner::new(&one, &two);
assert_eq!(scanner.find_ne_char_back(one.len(), two.len(), None), 3);
assert_eq!(scanner.find_ne_char_back(one.len(), two.len(), 2), 2);

pub fn find_ne_char<T>(
&mut self,
base_off: usize,
targ_off: usize,
stop: T
) -> usize where
T: Into<Option<usize>>,

Starting from the two provided offsets into the two ropes, returns the distance (in bytes) to the first non-equal codepoint. If no such position exists, returns the shortest distance to the end of a rope.

This can be thought of as the length of the longest common substring between base[base_off..] and target[targ_off..].

if stop is not None, the scan will stop at if it reaches this value.

Examples

let one = Rope::from("uh-oh🙈");
let two = Rope::from("uh-oh🙉");
let mut scanner = RopeScanner::new(&one, &two);
assert_eq!(scanner.find_ne_char(0, 0, None), 5);
assert_eq!(scanner.find_ne_char(0, 0, 3), 3);

pub fn find_min_diff_range(&mut self) -> (usize, usize)

Returns the positive offset from the start of the rope to the first non-equal byte, and the negative offset from the end of the rope to the first non-equal byte.

The two offsets are guaranteed not to overlap; thus sum(start_offset, end_offset) <= min(one.len(), two.len()).

Examples

let one = Rope::from("123xxx12345");
let two = Rope::from("123ZZZ12345");
let mut scanner = RopeScanner::new(&one, &two);
assert_eq!(scanner.find_min_diff_range(), (3, 5));

let one = Rope::from("friends");
let two = Rope::from("fiends");
let mut scanner = RopeScanner::new(&one, &two);
assert_eq!(scanner.find_min_diff_range(), (1, 5))