Struct xi_rope::delta::InsertDelta

pub struct InsertDelta<N: NodeInfo>(_);

A struct marking that a Delta contains only insertions. That is, it copies all of the old document in the same order. It has a Deref impl so all normal Delta methods can also be used on it.

Implementations

impl<N: NodeInfo> InsertDelta<N>

pub fn transform_expand(&self, xform: &Subset, after: bool) -> InsertDelta<N>

Do a coordinate transformation on an insert-only delta. The after parameter controls whether the insertions in self come after those specific in the coordinate transform.

pub fn transform_shrink(&self, xform: &Subset) -> InsertDelta<N>

Shrink a delta through a deletion of some of its copied regions with the same base. For example, if self applies to a union string, and xform is the deletions from that union, the resulting Delta will apply to the text.

pub fn inserted_subset(&self) -> Subset

Return a Subset containing the inserted ranges.

d.inserted_subset().delete_from_string(d.apply_to_string(s)) == s