Module xi_rope::multiset

A data structure for representing multi-subsets of sequences (typically strings).

Structs

Mapper

RangeIter

Subset

Represents a multi-subset of a string, that is a subset where elements can be included multiple times. This is represented as each element of the string having a “count” which is the number of times that element is included in the set.

SubsetBuilder

ZipIter

See Subset::zip

ZipSegment

See Subset::zip

Enums

CountMatcher

Determines which elements of a Subset a method applies to based on the count of the element.