Module xi_rope::rope

A rope data structure with a line count metric and (soon) other useful info.

Structs

BaseMetric

This metric let us walk utf8 text by code point.

ChunkIter

Lines

LinesMetric

LinesRaw

RopeInfo

Utf16CodeUnitsMetric

Functions

count_newlines

len_utf8_from_first_byte

Given the inital byte of a UTF-8 codepoint, returns the number of bytes required to represent the codepoint. RFC reference : https://tools.ietf.org/html/rfc3629#section-4

Type Aliases

Rope

A rope data structure.

RopeDelta

Represents a transform from one rope to another.

RopeDeltaElement

An element in a RopeDelta.