Struct xi_rope::tree::Cursor

pub struct Cursor<'a, N: 'a + NodeInfo> { /* fields omitted */ }

A data structure for traversing boundaries in a tree.

It is designed to be efficient both for random access and for iteration. The cursor itself is agnostic to which Metric is used to determine boundaries, but the methods to find boundaries are parametrized on the Metric.

A cursor can be valid or invalid. It is always valid when created or after set is called, and becomes invalid after prev or next fails to find a boundary.

Implementations

impl<'a> Cursor<'a, RopeInfo>

pub fn prev_codepoint(&mut self) -> Option<char>

Get previous codepoint before cursor position, and advance cursor backwards.

pub fn next_codepoint(&mut self) -> Option<char>

Get next codepoint after cursor position, and advance cursor.

pub fn peek_next_codepoint(&self) -> Option<char>

Get the next codepoint after the cursor position, without advancing the cursor.

pub fn next_grapheme(&mut self) -> Option<usize>

pub fn prev_grapheme(&mut self) -> Option<usize>

impl<'a, N: NodeInfo> Cursor<'a, N>

pub fn new(n: &'a Node<N>, position: usize) -> Cursor<'a, N>

Create a new cursor at the given position.

pub fn total_len(&self) -> usize

The length of the tree.

pub fn root(&self) -> &'a Node<N>

Return a reference to the root node of the tree.

pub fn get_leaf(&self) -> Option<(&'a N::L, usize)>

Get the current leaf of the cursor.

If the cursor is valid, returns the leaf containing the current position, and the offset of the current position within the leaf. That offset is equal to the leaf length only at the end, otherwise it is less than the leaf length.

pub fn set(&mut self, position: usize)

Set the position of the cursor.

The cursor is valid after this call.

Precondition: position is less than or equal to the length of the tree.

pub fn pos(&self) -> usize

Get the position of the cursor.

pub fn is_boundary<M: Metric<N>>(&mut self) -> bool

Determine whether the current position is a boundary.

Note: the beginning and end of the tree may or may not be boundaries, depending on the metric. If the metric is not can_fragment, then they always are.

pub fn prev<M: Metric<N>>(&mut self) -> Option<usize>

Moves the cursor to the previous boundary.

When there is no previous boundary, returns None and the cursor becomes invalid.

Return value: the position of the boundary, if it exists.

pub fn next<M: Metric<N>>(&mut self) -> Option<usize>

Moves the cursor to the next boundary.

When there is no next boundary, returns None and the cursor becomes invalid.

Return value: the position of the boundary, if it exists.

pub fn at_or_next<M: Metric<N>>(&mut self) -> Option<usize>

Returns the current position if it is a boundary in this Metric, else behaves like next.

pub fn at_or_prev<M: Metric<N>>(&mut self) -> Option<usize>

Returns the current position if it is a boundary in this Metric, else behaves like prev.

pub fn iter<'c, M: Metric<N>>(&'c mut self) -> CursorIter<'c, 'a, N, M>

Returns an iterator with this cursor over the given Metric.

Examples:

let text: Rope = "one line\ntwo line\nred line\nblue".into();
let mut cursor = Cursor::new(&text, 0);
let line_offsets = cursor.iter::<LinesMetric>().collect::<Vec<_>>();
assert_eq!(line_offsets, vec![9, 18, 27]);

pub fn next_leaf(&mut self) -> Option<(&'a N::L, usize)>

Move to beginning of next leaf.

Return value: same as get_leaf.

pub fn prev_leaf(&mut self) -> Option<(&'a N::L, usize)>

Move to beginning of previous leaf.

Return value: same as get_leaf.