Trait xi_rope::tree::Leaf

pub trait Leaf: Sized + Clone + Default {
    fn len(&self) -> usize;

    fn is_ok_child(&self) -> bool;

    fn push_maybe_split(&mut self, other: &Self, iv: Interval) -> Option<Self>;

    fn subseq(&self, iv: Interval) -> Self { ... }
}

A trait for the leaves of trees of type Node.

Two leafs can be concatenated using push_maybe_split.

Required methods

fn len(&self) -> usize

Measurement of leaf in base units. A ‘base unit’ refers to the smallest discrete unit by which a given concrete type can be indexed. Concretely, for Rust’s String type the base unit is the byte.

fn is_ok_child(&self) -> bool

Generally a minimum size requirement for leaves.

fn push_maybe_split(&mut self, other: &Self, iv: Interval) -> Option<Self>

Combine the part other denoted by the Interval iv into self, optionly splitting off a new Leaf if self would have become too big. Returns either None if no splitting was needed, or Some(rest) if rest was split off.

Interval is in “base units”. Generally implements a maximum size.

Invariants:

If one or the other input is empty, then no split.

If either input satisfies is_ok_child, then, on return, self satisfies this, as does the optional split.

Provided methods

fn subseq(&self, iv: Interval) -> Self

Same meaning as push_maybe_split starting from an empty leaf, but maybe can be implemented more efficiently?