Crate xi_rope

Trees for text.

Re-exports

pub use crate::delta::Builder as DeltaBuilder;

pub use crate::delta::Delta;

pub use crate::delta::DeltaElement;

pub use crate::delta::Transformer;

pub use crate::interval::Interval;

pub use crate::rope::LinesMetric;

pub use crate::rope::Rope;

pub use crate::rope::RopeDelta;

pub use crate::rope::RopeInfo;

pub use crate::tree::Cursor;

pub use crate::tree::Metric;

Modules

breaks

A module for representing a set of breaks, typically used for storing the result of line breaking.

compare

Fast comparison of rope regions, principally for diffing.

delta

A data structure for representing editing operations on ropes. It’s useful to explicitly represent these operations so they can be shared across multiple subsystems.

diff

Computing deltas between two ropes.

engine

An engine for handling edits (possibly from async sources) and undo. It conceptually represents the current text and all edit history for that text.

find

Implementation of string finding in ropes.

interval

Closed-open intervals, and operations on them.

multiset

A data structure for representing multi-subsets of sequences (typically strings).

rope

A rope data structure with a line count metric and (soon) other useful info.

spans

A module for representing spans (in an interval tree), useful for rich text annotations. It is parameterized over a data type, so can be used for storing different annotations.

tree

A general b-tree structure suitable for ropes and the like.